Green-32355: Update gems
The older versions of the ffi gem do not compile properly on newer systems. By updating the gems (`bundle update`) we will begin to use more recent versions of the required gems. This will not bring all the gems up to the latest but update within the requirements already defined the the Gemfile.